+++ Game Server/Maps/MapPacket.cs
@@ -2,6 +2,7 @@
 using Lidgren.Network;
 using GameServer.Network;
 using GameServer.Server;
+using GameServer.Player;
 
 namespace GameServer.Maps {
     public partial class Map {
@@ -13,7 +14,7 @@ namespace GameServer.Maps {
             var npcData = MapManager.FindMapByID(pData.WorldID).Npcs;
 
             var buffer = GameNetwork.CreateMessage();
-            buffer.Write((int)PacketList.GameServer_SendNpc);
+            buffer.Write((short)PacketList.GameServer_SendNpc);
             buffer.Write(npcData.Count);
 
             foreach(var npc in npcData) {
@@ -36,7 +37,7 @@ namespace GameServer.Maps {
         /// <param name="npc"></param>
         public void SendNpcMove(MapNpcData npc) {
             var buffer = GameNetwork.CreateMessage();
-            buffer.Write((int)PacketList.GameServer_NpcMove);
+            buffer.Write((short)PacketList.GameServer_NpcMove);
             buffer.Write(npc.UniqueID);
             buffer.Write((byte)npc.Direction);
 
@@ -66,7 +67,7 @@ namespace GameServer.Maps {
 
         public static void SendPlayerMapMove(NetConnection connection, int playerID, byte direction) {
             var buffer = GameNetwork.CreateMessage();
-            buffer.Write((int)PacketList.GameServer_Client_PlayerMapMove);
+            buffer.Write((short)PacketList.GameServer_Client_PlayerMapMove);
             buffer.Write(playerID);
             buffer.Write(direction);
 
@@ -75,7 +76,7 @@ namespace GameServer.Maps {
 
         public static void SendMapPlayer(NetConnection connection, int playerID, string name, short sprite, byte direction, short x, short y) {
             var buffer = GameNetwork.CreateMessage();
-            buffer.Write((int)PacketList.GameServer_Client_GetMapPlayer);
+            buffer.Write((short)PacketList.GameServer_Client_GetMapPlayer);
             buffer.Write(playerID);
             buffer.Write(name);
             buffer.Write(sprite);
@@ -89,7 +90,7 @@ namespace GameServer.Maps {
 
         public static void RemovePlayerOnMap(NetConnection connection, int playerID) {
             var buffer = GameNetwork.CreateMessage();
-            buffer.Write((int)PacketList.GameServer_Client_RemovePlayerFromMap);
+            buffer.Write((short)PacketList.GameServer_Client_RemovePlayerFromMap);
             buffer.Write(playerID);
 
             GameNetwork.SendDataTo(connection, buffer, NetDeliveryMethod.ReliableOrdered);
